Move struct `string_port` definition into new header `string_port.hpp`
