+++ include/meevax/kernel/port.hpp
@@ -61,7 +61,7 @@ inline namespace kernel
 
     auto read() -> object;
 
-    virtual operator std::string();
+    explicit virtual operator std::string();
   };
 
   struct textual_output_port : public virtual textual_port, public virtual output_port
@@ -95,24 +95,6 @@ inline namespace kernel
   };
 
   auto operator <<(std::ostream &, standard_error_port const&) -> std::ostream &;
-
-  struct string_port : public textual_input_port, public textual_output_port
-  {
-    std::stringstream stringstream;
-
-    template <typename... Ts>
-    explicit string_port(Ts&&... xs)
-      : stringstream { std::forward<decltype(xs)>(xs)... }
-    {}
-
-    operator std::istream &() override;
-
-    operator std::ostream &() override;
-
-    operator std::string() override;
-  };
-
-  auto operator <<(std::ostream &, string_port const&) -> std::ostream &;
 } // namespace kernel
 } // namespace meevax
