+++ src/kernel/library.cpp
@@ -22,6 +22,7 @@
 #include <meevax/kernel/disassemble.hpp>
 #include <meevax/kernel/file_port.hpp>
 #include <meevax/kernel/library.hpp>
+#include <meevax/kernel/string_port.hpp>
 
 namespace meevax
 {
